+++ weblate/settings_docker.py
@@ -874,6 +874,7 @@ CHECK_LIST = [
     "weblate.checks.format.JavaMessageFormatCheck",
     "weblate.checks.format.PercentPlaceholdersCheck",
     "weblate.checks.format.I18NextInterpolationCheck",
+    "weblate.checks.format.ESTemplateLiteralsCheck",
     "weblate.checks.angularjs.AngularJSInterpolationCheck",
     "weblate.checks.qt.QtFormatCheck",
     "weblate.checks.qt.QtPluralCheck",