+++ docs/changes.rst
@@ -20,6 +20,7 @@ Not yet released.
 * Added support for ModernMT.
 * Allow to avoid overwriting approved translations on file upload.
 * Dropped support for some compatibility URL redirects.
+* Added check for ECMAScript template literals.
 
 Weblate 4.1.1
 -------------